requirements.txt
openai-agents
openai-agents[voice]
python-dotenv
sounddevice
fastapi
uvicorn
pydantic
numpy